usb_modeswitch.sh (udev): simplify the script's ending (rewrite equivalently)
If there is only one "/" in $1, $p2 would be computed by the old code to
the same value as ${1##*/}. Thus, the code rewritten this way is
equivalent to the old code.

If there are many "/", then we don't want to leave any in the argument
for usb_modeswitch_dispatcher, because it can fail handling it anyway
(the first failure being the failure to create the log file if logging
is on).

(As I see, the last component of the device path must be a complete
designation of the device.)