checkSuccess(): also define other local vars with an initializer
Perhaps, this breaks the style, but declaring a variable just in the
code block where it is actually used (and with initializer) makes the
code easier to read and spot errors (and not suspect false ones).

As for newTargetCount: sure, it must be assigned by a function call,
but even if that function fails (to assign it) for some reason,
having it initialized to 0 must be safe, because then the condition
(newTargetCount > targetDeviceCount) still can't become true.