usb_modeswitch.c: always null-terminate the strings with hex messages
Since these strings with hexadecimal messages become arguments to
strlen and strcpy, they always need to be null-terminated.

I didn't want to change the size of ByteString (the bytes represented
by these hexadecimal messages), which was LINE_DIM/2, and since
ByteString needs not be null-terminated (sent or received via
libusb_bulk_transfer), the size of the buffer for the null-terminated string
holding the hexadecimal representation of ByteString was changed to
LINE_DIM+1 by me.

The fixed issue is the following:

A maximally long message could be written to these buffers (either
from the command-line options/config or perhaps from USB). It wouldn't
be null-terminated. A buffer overrun would happen when reading it,
and, ultimately, a buffer overflow would happen when writing to such a
buffer (copying the messages with strcpy). Leading to overwriting
other memory.

Perhaps, this could be not too serious, if command-line/config options
couldn't be longer than 1024 (LINE_DIM), because they are
null-terminated. The data read from USB doesn't seem to be actually
converted to a hexadecimal representation and stored as a message.
At least, this would fix compiler warnings, if the issue were not serious.

Reported by GCC:

In file included from /usr/include/string.h:532,
                 from usb_modeswitch.c:52:
In function 'strncpy',
    inlined from 'readArguments' at usb_modeswitch.c:367:14:
/usr/include/bits/string_fortified.h:91:10: warning: '__builtin_strncpy' specified bound 1024 equals destination size [-Wstringop-truncation]
   91 |   return __builtin___strncpy_chk (__dest, __src, __len, __bos (__dest));
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In function 'strncpy',
    inlined from 'readArguments' at usb_modeswitch.c:369:14:
/usr/include/bits/string_fortified.h:91:10: warning: '__builtin_strncpy' specified bound 1024 equals destination size [-Wstringop-truncation]
   91 |   return __builtin___strncpy_chk (__dest, __src, __len, __bos (__dest));
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In function 'strncpy',
    inlined from 'readArguments' at usb_modeswitch.c:368:14:
/usr/include/bits/string_fortified.h:91:10: warning: '__builtin_strncpy' specified bound 1024 equals destination size [-Wstringop-truncation]
   91 |   return __builtin___strncpy_chk (__dest, __src, __len, __bos (__dest));
      |          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~