*-checkinstall pkgs to check compatibility with sysvinit or systemd
There can be extra autodeps from the scripts, which are actually
optional, and would make it conflict with the other startup system.
(E.g., systemctl. However, the installation of systemd-utils is
actually possible together with sysvinit currently in Sisyphus, p9,
p8. Still there can be a wish not to have this optional dependency.)